refactor: deslop Slack code channels
