+++ agent/tools/open_pull_request.py
@@ -18,7 +18,7 @@ from ..utils.slack_code_channels import (
     is_code_channel_session,
     repo_context_bar_items,
     set_context_bar,
-    set_diff_view,
+    set_view,
 )
 
 logger = logging.getLogger(__name__)
@@ -526,46 +526,6 @@ async def _thread_pull_requests(thread_id: str) -> list[dict[str, Any]]:
     ]
 
 
-async def _publish_code_channel_pr(
-    client: httpx.AsyncClient,
-    token: str,
-    owner: str,
-    repo: str,
-    pr_number: int,
-    *,
-    pr_url: str,
-    head: str,
-    base: str,
-    configurable: dict[str, Any],
-) -> None:
-    """Mirror a new PR into the session's Slack code channel chrome."""
-    thread_id = configurable.get("thread_id")
-    slack_thread = configurable.get("slack_thread")
-    active = await get_active_slack_thread(
-        get_client(),
-        thread_id if isinstance(thread_id, str) else None,
-        slack_thread if isinstance(slack_thread, dict) else None,
-    )
-    if not active or not is_code_channel_session(str(active.get("thread_ts") or "")):
-        return
-    channel_id = str(active.get("channel_id") or "")
-    await set_context_bar(
-        channel_id,
-        repo_context_bar_items({"owner": owner, "name": repo}, branch=head, pr_url=pr_url),
-    )
-    resp = await client.get(
-        f"{GITHUB_API}/repos/{owner}/{repo}/pulls/{pr_number}",
-        headers={**_auth_headers(token), "Accept": "application/vnd.github.v3.diff"},
-    )
-    if resp.status_code == 200 and resp.text.strip():
-        await set_diff_view(
-            channel_id,
-            resp.text,
-            base_branch=base,
-            head_branch=head,
-        )
-
-
 async def _record_pr_telemetry(
     *,
     client: httpx.AsyncClient,
@@ -676,17 +636,34 @@ async def _record_pr_telemetry(
             if repo_private is not None:
                 metadata["repo_private"] = repo_private
             await get_client().threads.update(thread_id=thread_id, metadata=metadata)
-            await _publish_code_channel_pr(
-                client,
-                token,
-                owner,
-                repo,
-                pr_number,
-                pr_url=pr_url if isinstance(pr_url, str) else "",
-                head=head,
-                base=base,
-                configurable=configurable,
+            slack_thread = configurable.get("slack_thread")
+            active = await get_active_slack_thread(
+                get_client(),
+                thread_id,
+                slack_thread if isinstance(slack_thread, dict) else None,
             )
+            if active and is_code_channel_session(str(active.get("thread_ts") or "")):
+                channel_id = str(active.get("channel_id") or "")
+                await set_context_bar(
+                    channel_id,
+                    repo_context_bar_items(
+                        {"owner": owner, "name": repo},
+                        branch=head,
+                        pr_url=pr_url if isinstance(pr_url, str) else "",
+                    ),
+                )
+                response = await client.get(
+                    f"{GITHUB_API}/repos/{owner}/{repo}/pulls/{pr_number}",
+                    headers={**_auth_headers(token), "Accept": "application/vnd.github.v3.diff"},
+                )
+                if response.status_code == 200 and response.text.strip():
+                    await set_view(
+                        channel_id,
+                        "diff",
+                        content=response.text,
+                        base_branch=base,
+                        head_branch=head,
+                    )
     except Exception:
         logger.debug(
             "Failed to record PR usage for %s/%s#%s", owner, repo, pr_number, exc_info=True
